+++ VoxelGame/Logic/Blocks/Core/CustomModelBlock.cs
@@ -10,7 +10,7 @@ namespace VoxelGame.Logic.Blocks
 {
     /// <summary>
     /// A block that loads its complete model from a file. The block can only be placed on top of solid and full blocks.
-    /// Data bit usage: <c>-----</c>
+    /// Data bit usage: <c>------</c>
     /// </summary>
     public class CustomModelBlock : Block
     {
@@ -48,7 +48,7 @@ namespace VoxelGame.Logic.Blocks
             vertCount = (uint)(blockModel.VertexCount);
         }
 
-        public override uint GetMesh(BlockSide side, byte data, out float[] vertices, out int[] textureIndices, out uint[] indices, out TintColor tint, out bool isAnimated)
+        public override uint GetMesh(BlockSide side, uint data, out float[] vertices, out int[] textureIndices, out uint[] indices, out TintColor tint, out bool isAnimated)
         {
             vertices = this.vertices;
             textureIndices = texIndices;
@@ -74,7 +74,7 @@ namespace VoxelGame.Logic.Blocks
             }
         }
 
-        internal override void BlockUpdate(int x, int y, int z, byte data, BlockSide side)
+        internal override void BlockUpdate(int x, int y, int z, uint data, BlockSide side)
         {
             if (side == BlockSide.Bottom && !(Game.World.GetBlock(x, y - 1, z, out _) ?? Block.Air).IsSolidAndFull)
             {
